feat: remove @zarrita/v3
