+++ packages/core/src/index.ts
@@ -1,15 +1,9 @@
 export * as ops from "./ops.js";
 export { KeyError, NodeNotFoundError } from "./errors.js";
-export { Array, create, Group, Location, open, root } from "./hierarchy.js";
-export {
-	is_dtype,
-	json_decode_object,
-	json_encode_object,
-	range,
-	slice,
-} from "./util.js";
-export { registry } from "./codec-registry.js";
-export { get as get_with_setter } from "./get.js";
-export { set as set_with_setter } from "./set.js";
+export { Array, Group, Location, root } from "./hierarchy.js";
+export { open } from "./open.js";
+export { create } from "./create.js";
+export { registry } from "./codecs.js";
+export { slice } from "./util.js";
 export type * from "./types.js";
-export type { Codec } from "numcodecs";
+export type * from "./metadata.js";
