+++ packages/core/src/indexing.ts
@@ -1,6 +1,6 @@
+import type { Indices, Slice } from "./types.js";
 import { IndexError } from "./errors.js";
 import { product, range, slice } from "./util.js";
-import type { Indices, Slice } from "./types.js";
 
 function err_too_many_indices(
 	selection: (number | Slice)[],
