--- packages/v3/__tests__/index.test.ts
@@ -1,312 +1,0 @@
-import { test, expect } from "vitest";
-import ndarray from "ndarray";
-
-import MemStore from "@zarrita/storage/mem";
-
-import { ops, json_decode_object, range } from "@zarrita/core";
-let { get, set } = ops;
-
-import * as zarr from "../index.js";
-
-test("create root group", async () => {
-	let h = await zarr.create_hierarchy(new MemStore());
-	let attrs = { hello: "world" };
-	let grp = await zarr.create_group(h, "/", { attrs });
-	expect(grp.path).toBe("/");
-	expect(await grp.attrs()).toStrictEqual(attrs);
-	expect(h.store.has("/zarr.json")).true;
-	expect(json_decode_object(h.store.get("/zarr.json")!)).toStrictEqual(
-		{
-			zarr_format: "https://purl.org/zarr/spec/protocol/core/3.0",
-			metadata_encoding: "https://purl.org/zarr/spec/protocol/core/3.0",
-			metadata_key_suffix: ".json",
-			extensions: [],
-		},
-	);
-	expect(h.store.has("/meta/root.group.json")).true;
-	expect(json_decode_object(h.store.get("/meta/root.group.json")!)).toStrictEqual(
-		{
-			extensions: [],
-			attributes: attrs,
-		},
-	);
-});
-
-test("create array", async () => {
-	let h = await zarr.create_hierarchy(new MemStore());
-	let attrs = { question: "life", answer: 42 };
-	let a = await zarr.create_array(h, "/arthur/dent", {
-		shape: [5, 10],
-		dtype: "<i4",
-		chunk_shape: [2, 5],
-		attrs,
-	});
-	expect(a.path).toBe("/arthur/dent");
-	expect(a.name).toBe("dent");
-	expect(a.ndim).toBe(2);
-	expect(a.shape).toStrictEqual([5, 10]);
-	expect(a.dtype).toBe("<i4");
-	expect(a.chunk_shape).toStrictEqual([2, 5]);
-	expect(await a.attrs()).toStrictEqual(attrs);
-
-	expect(h.store.has("/meta/root/arthur/dent.array.json")).true;
-	expect(json_decode_object(h.store.get("/meta/root/arthur/dent.array.json")!)).toStrictEqual(
-		{
-			shape: [5, 10],
-			data_type: "<i4",
-			chunk_grid: {
-				type: "regular",
-				chunk_shape: [2, 5],
-				separator: "/",
-			},
-			chunk_memory_layout: "C",
-			fill_value: null,
-			extensions: [],
-			attributes: attrs,
-		},
-	);
-});
-
-test("create group from another group", async () => {
-	let h = await zarr.create_hierarchy(new Map());
-	let grp = await zarr.create_group(h, "/tricia/mcmillan");
-	await zarr.create_group(grp, "relative");
-	await zarr.create_group(grp, "/absolute");
-	expect(new Set(h.store.keys())).toStrictEqual(new Set([
-			"/zarr.json",
-			"/meta/root/tricia/mcmillan.group.json",
-			"/meta/root/tricia/mcmillan/relative.group.json",
-			"/meta/root/absolute.group.json",
-		]));
-});
-
-test("create explicit group and access implicit group", async () => {
-	let h = await zarr.create_hierarchy(new MemStore());
-	await zarr.create_group(h, "/tricia/mcmillan");
-	let grp = await zarr.get_implicit_group(h, "/tricia");
-	expect(grp).toBeInstanceOf(zarr.ImplicitGroup);
-});
-
-test("create nodes via groups", async () => {
-	let h = await zarr.create_hierarchy(new MemStore());
-	let marvin = await zarr.create_group(h, "/marvin");
-	let paranoid = await zarr.create_group(marvin, "paranoid");
-	let android = await zarr.create_array(marvin, "android", {
-		shape: [42, 42],
-		dtype: "|u1",
-		chunk_shape: [2, 2],
-	});
-	expect(marvin).toBeInstanceOf(zarr.ExplicitGroup);
-	expect(marvin.path).toBe("/marvin");
-	expect(paranoid).toBeInstanceOf(zarr.ExplicitGroup);
-	expect(paranoid.path).toBe("/marvin/paranoid");
-	expect(android).toBeInstanceOf(zarr.Array);
-	expect(android.path).toBe("/marvin/android");
-});
-
-test("get_children", async () => {
-	let h = await zarr.create_hierarchy(new MemStore());
-
-	await zarr.create_array(h, "/arthur/dent", {
-		shape: [5, 10],
-		dtype: "<i4",
-		chunk_shape: [2, 5],
-		attrs: { question: "life", answer: 42 },
-	});
-
-	await zarr.create_array(h, "/deep/thought", {
-		shape: [7500000],
-		dtype: ">f8",
-		chunk_shape: [42],
-	});
-
-	await zarr.create_group(h, "/tricia/mcmillan");
-
-	let marvin = await zarr.create_group(h, "/marvin");
-	await zarr.create_group(marvin, "paranoid");
-	await zarr.create_array(marvin, "android", {
-		shape: [42, 42],
-		dtype: "|u1",
-		chunk_shape: [2, 2],
-	});
-
-	expect(await zarr.get_children(h, "/")).toStrictEqual(new Map([
-			["arthur", "implicit_group"],
-			["deep", "implicit_group"],
-			["marvin", "explicit_group"],
-			["tricia", "implicit_group"],
-		]));
-
-	expect(await zarr.get_children(h, "/tricia")).toStrictEqual(
-		new Map().set("mcmillan", "explicit_group"),
-	);
-
-	expect(await zarr.get_children(h, "/arthur")).toStrictEqual(
-		new Map().set("dent", "array"),
-	);
-});
-
-test("get_nodes", async () => {
-	let h = await zarr.create_hierarchy(new MemStore());
-
-	await zarr.create_array(h, "/arthur/dent", {
-		shape: [5, 10],
-		dtype: "<i4",
-		chunk_shape: [2, 5],
-		attrs: { question: "life", answer: 42 },
-	});
-
-	await zarr.create_array(h, "/deep/thought", {
-		shape: [7500000],
-		dtype: ">f8",
-		chunk_shape: [42],
-	});
-
-	await zarr.create_group(h, "/tricia/mcmillan");
-
-	let marvin = await zarr.create_group(h, "/marvin");
-	await zarr.create_group(marvin, "paranoid");
-	await zarr.create_array(marvin, "android", {
-		shape: [42, 42],
-		dtype: "|u1",
-		chunk_shape: [2, 2],
-	});
-
-	expect(await zarr.get_children(h, "/")).toStrictEqual(new Map([
-			["arthur", "implicit_group"],
-			["deep", "implicit_group"],
-			["marvin", "explicit_group"],
-			["tricia", "implicit_group"],
-		]));
-
-	expect(await zarr.get_children(h, "/tricia")).toStrictEqual(
-		new Map().set("mcmillan", "explicit_group"),
-	);
-
-	expect(await zarr.get_children(h, "/arthur")).toStrictEqual(
-		new Map().set("dent", "array"),
-	);
-
-	expect(await zarr.get_nodes(h)).toStrictEqual(new Map([
-			["/", "implicit_group"],
-			["/arthur", "implicit_group"],
-			["/arthur/dent", "array"],
-			["/deep", "implicit_group"],
-			["/deep/thought", "array"],
-			["/marvin", "explicit_group"],
-			["/marvin/android", "array"],
-			["/marvin/paranoid", "explicit_group"],
-			["/tricia", "implicit_group"],
-			["/tricia/mcmillan", "explicit_group"],
-		]));
-});
-
-test("Read and write array data - builtin", async () => {
-	let h = await zarr.create_hierarchy(new MemStore());
-	let a = await zarr.create_array(h, "/arthur/dent", {
-		shape: [5, 10],
-		dtype: "<i4",
-		chunk_shape: [2, 5],
-	});
-
-	let res = await get(a, [null, null]);
-	expect(res.shape).toStrictEqual([5, 10]);
-	expect(res.data).toStrictEqual(new Int32Array(50));
-
-	res = await get(a);
-	expect(res.shape).toStrictEqual([5, 10]);
-	expect(res.data).toStrictEqual(new Int32Array(50));
-
-	await set(a, [0, null], 42);
-	expect((await get(a, null)).data).toStrictEqual(
-		new Int32Array(50).fill(42, 0, 10),
-	);
-
-	const expected = new Int32Array(50).fill(42, 0, 10);
-	[10, 20, 30, 40].forEach((i) => (expected[i] = 42));
-	await set(a, [null, 0], 42);
-	expect((await get(a, null)).data).toStrictEqual(expected);
-
-	await set(a, null, 42);
-	expected.fill(42);
-	expect((await get(a, null)).data).toStrictEqual(expected);
-
-	let arr = ndarray(new Int32Array([...Array(10).keys()]), [10]);
-	expected.set(arr.data);
-	await set(a, [0, null], arr);
-	expect((await get(a, null)).data).toStrictEqual(expected);
-
-	arr = ndarray(new Int32Array(range(50)), [5, 10]);
-	await set(a, null, arr);
-	expect((await get(a, null)).data).toStrictEqual(arr.data);
-
-	// Read array slices
-	res = await get(a, [null, 0]);
-	expect(res.shape).toStrictEqual([5]);
-	expect(res.data).toStrictEqual(new Int32Array([0, 10, 20, 30, 40]));
-
-	res = await get(a, [null, 1]);
-	expect(res.shape).toStrictEqual([5]);
-	expect(res.data).toStrictEqual(new Int32Array([1, 11, 21, 31, 41]));
-
-	res = await get(a, [0, null]);
-	expect(res.shape).toStrictEqual([10]);
-	expect(res.data).toStrictEqual(new Int32Array([0, 1, 2, 3, 4, 5, 6, 7, 8, 9]));
-
-	res = await get(a, [1, null]);
-	expect(res.shape).toStrictEqual([10]);
-	expect(res.data).toStrictEqual(new Int32Array([10, 11, 12, 13, 14, 15, 16, 17, 18, 19]));
-
-	res = await get(a, [null, zarr.slice(0, 7)]);
-	expect(res.shape).toStrictEqual([5, 7]);
-	// deno-fmt-ignore
-	expect(res.data).toStrictEqual(new Int32Array([ 
-		 0,  1,  2,  3,  4,
-		 5,  6, 10, 11, 12,
-		13, 14, 15, 16, 20,
-		21, 22, 23, 24, 25,
-		26, 30, 31, 32, 33,
-		34, 35, 36, 40, 41,
-		42, 43, 44, 45, 46,
-	]));
-
-	res = await get(a, [zarr.slice(0, 3), null]);
-	expect(res.shape).toStrictEqual([3, 10]);
-	// deno-fmt-ignore
-	expect(res.data).toStrictEqual(new Int32Array([
-		 0,  1,  2,  3,  4,  5,  6,  7,  8,  9,
-		10, 11, 12, 13, 14, 15, 16, 17, 18, 19,
-		20, 21, 22, 23, 24, 25, 26, 27, 28, 29,
-	]));
-	res = await get(a, [zarr.slice(0, 3), zarr.slice(0, 7)]);
-	expect(res.shape).toStrictEqual([3, 7]);
-	// deno-fmt-ignore
-	expect(res.data).toStrictEqual(new Int32Array([
-		 0,  1,  2,  3,  4,  5,  6,
-		10, 11, 12, 13, 14, 15, 16,
-		20, 21, 22, 23, 24, 25, 26,
-	]));
-
-	res = await get(a, [zarr.slice(1, 4), zarr.slice(2, 7)]);
-	expect(res.shape).toStrictEqual([3, 5]);
-	// deno-fmt-ignore
-	expect(res.data).toStrictEqual(new Int32Array([
-		12, 13, 14, 15, 16,
-		22, 23, 24, 25, 26,
-		32, 33, 34, 35, 36,
-	]));
-
-	let b = await zarr.create_array(h, "/deep/thought", {
-		shape: [7500000],
-		dtype: ">f8",
-		chunk_shape: [42],
-	});
-
-	let resb = await get(b, [zarr.slice(10)]);
-	expect(resb.shape).toStrictEqual([10]);
-	expect(resb.data).toStrictEqual(new Float64Array(10));
-
-	expected.fill(1, 0, 5);
-	await set(b, [zarr.slice(5)], 1);
-	expect((await get(b, [zarr.slice(10)])).data).toStrictEqual(new Float64Array(10).fill(1, 0, 5));
-});
