--- packages/v3/package.json
@@ -1,24 +1,0 @@
-{
-  "name": "@zarrita/v3",
-  "type": "module",
-  "version": "0.0.0",
-  "scripts": {
-    "build": "tsc",
-    "clean": "rm -rf dist"
-  },
-  "exports": {
-    ".": {
-      "types": "./dist/index.d.ts",
-      "import": "./dist/index.js"
-    }
-  },
-  "dependencies": {
-    "@zarrita/core": "workspace:^"
-  },
-  "devDependencies": {
-    "@types/ndarray": "^1.0.11",
-    "@zarrita/storage": "workspace:^",
-    "@zarrita/typedarray": "workspace:^",
-    "ndarray": "^1.0.19"
-  }
-}
