--- packages/v3/tsconfig.json
@@ -1,7 +1,0 @@
-{
-  "extends": "../../tsconfig.json",
-  "compilerOptions": {
-    "outDir": "dist"
-  },
-  "include": ["index.ts", "src/**/*.ts", "__tests__/**/*.ts"]
-}
